Fix getElementWidth function to use parseFloat for consistent calculations

+++ js/task-3.js
@@ -1,6 +1,6 @@
 'use strict'
 function getElementWidth(content, padding, border) {
-    return Number.parseInt(content) + 2*Number.parseInt(padding) + 2*Number.parseFloat(border);
+    return Number.parseFloat(content) + 2*Number.parseFloat(padding) + 2*Number.parseFloat(border);
 }
 
 console.log(getElementWidth("50px", "8px", "4px")); // 74
